Rename some CDS-related classes.
